Created New Comment Components and Uncommented TestCases, fixed Delete Functionality by adding bidirectional component in Image
