+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -1,9 +1,9 @@
 package ImageHoster.controller;
 
-import ImageHoster.model.Comment;
 import ImageHoster.model.Image;
 import ImageHoster.model.Tag;
 import ImageHoster.model.User;
+import ImageHoster.service.CommentService;
 import ImageHoster.service.ImageService;
 import ImageHoster.service.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,10 +17,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
 import java.util.*;
 
 @Controller
@@ -31,6 +27,9 @@ public class ImageController {
 
     @Autowired
     private TagService tagService;
+
+    @Autowired
+    private CommentService commentService;
 
     //This method displays all the images in the user home page after successful login
     @RequestMapping("images")
@@ -55,7 +54,7 @@ public class ImageController {
         Image image = imageService.getImage(imageId);
         model.addAttribute("image", image);
         model.addAttribute("tags", image.getTags());
-        model.addAttribute("comments", imageService.getAllCommentsByImage(image));
+        model.addAttribute("comments", commentService.getAllCommentsByImage(image));
         return "images/image";
     }
 
@@ -111,7 +110,7 @@ public class ImageController {
             String error = "Only the owner of the image can edit the image";
             model.addAttribute("editError",error);
             model.addAttribute("tags", image.getTags());
-            model.addAttribute("comments", imageService.getAllCommentsByImage(image));
+            model.addAttribute("comments", commentService.getAllCommentsByImage(image));
             return "images/image";
         }
     }
@@ -167,26 +166,9 @@ public class ImageController {
             model.addAttribute("image", image);
             model.addAttribute("tags", image.getTags());
             model.addAttribute("deleteError",error);
-            model.addAttribute("comments", imageService.getAllCommentsByImage(image));
+            model.addAttribute("comments", commentService.getAllCommentsByImage(image));
             return "images/image";
         }
-    }
-
-    @RequestMapping(value = "/image/{id}/{title}/comments", method = RequestMethod.POST)
-    public String addComment(@PathVariable("id") Integer id, @PathVariable("title") String title, @RequestParam("comment") String comment, HttpSession httpSession){
-        Comment newComment = new Comment();
-        Date input = new Date();
-        Instant instant = input.toInstant();
-        ZonedDateTime zdt = instant.atZone(ZoneId.systemDefault());
-        LocalDate date = zdt.toLocalDate();
-        User user = (User) httpSession.getAttribute("loggeduser");
-        Image image = imageService.getImage(id);
-        newComment.setImage(image);
-        newComment.setUser(user);
-        newComment.setCreatedDate(date);
-        newComment.setText(comment);
-        imageService.addCommentToImage(newComment);
-        return "redirect:/images/" + id + '/' + title;
     }
 
     //This method converts the image to Base64 format
