+++ src/main/java/ImageHoster/repository/CommentRepository.java
@@ -1,0 +1,36 @@
+package ImageHoster.repository;
+
+import ImageHoster.model.Comment;
+import org.springframework.stereotype.Repository;
+
+import javax.persistence.*;
+import java.util.List;
+
+@Repository
+public class CommentRepository {
+    @PersistenceUnit(unitName = "imageHoster")
+    private EntityManagerFactory emf;
+
+    public List<Comment> getAllComments(Integer imageId) {
+        EntityManager em = emf.createEntityManager();
+        TypedQuery<Comment> query = em.createQuery("SELECT c FROM Comment c WHERE c.image.id = :imageId", Comment.class);
+        query.setParameter("imageId",imageId);
+        List<Comment> resultList = query.getResultList();
+
+        return resultList;
+    }
+
+    public void addComment(Comment newComment){
+        EntityManager em = emf.createEntityManager();
+        EntityTransaction transaction = em.getTransaction();
+
+        try {
+            transaction.begin();
+            //persist() method changes the state of the model object from transient state to persistence state
+            em.persist(newComment);
+            transaction.commit();
+        } catch (Exception e) {
+            transaction.rollback();
+        }
+    }
+}
