+++ src/main/java/ImageHoster/repository/ImageRepository.java
@@ -111,27 +111,6 @@ public class ImageRepository {
         }
     }
 
-    public List<Comment> getAllComments(Integer imageId) {
-        EntityManager em = emf.createEntityManager();
-        TypedQuery<Comment> query = em.createQuery("SELECT c FROM Comment c WHERE c.image.id = :imageId", Comment.class);
-        query.setParameter("imageId",imageId);
-        List<Comment> resultList = query.getResultList();
 
-        return resultList;
-    }
-
-    public void addComment(Comment newComment){
-        EntityManager em = emf.createEntityManager();
-        EntityTransaction transaction = em.getTransaction();
-
-        try {
-            transaction.begin();
-            //persist() method changes the state of the model object from transient state to persistence state
-            em.persist(newComment);
-            transaction.commit();
-        } catch (Exception e) {
-            transaction.rollback();
-        }
-    }
 
 }
