+++ src/main/java/ImageHoster/service/ImageService.java
@@ -54,11 +54,4 @@ public class ImageService {
         return user.getId().equals(image.getUser().getId());
     }
 
-    public List<Comment> getAllCommentsByImage(Image image){
-        return imageRepository.getAllComments(image.getId());
-    }
-
-    public void addCommentToImage(Comment newComment){
-        imageRepository.addComment(newComment);
-    }
 }
