+++ src/test/java/ImageHoster/controller/CommentControllerTest.java
@@ -1,4 +1,3 @@
-/*
 package ImageHoster.controller;
 
 
@@ -65,4 +64,4 @@ public class CommentControllerTest {
                 .andExpect(redirectedUrl("/images/1/new"));
     }
 }
-*/
+
